Ensure spec files are up to date before executing (#57)
Whenever there's a bug or unclear test message that we want to update after students have already started we can't affect students spec files (unless we do something fancy like write a script that re-writes their spec file and have them run a command that downloads and runs it)

This commit enhances rails grade check to see if the local spec folder/files are up to date with the latest version stored on appdev-projects/upstream version. Any unstaged/untracked files are discarded first, but if they do match, the normal rails grade thing happens. If they don't match, the contents of the spec folder are downloaded from upstream and then the rspec is run and the results are submitted to Grades.

This way, students can still just run rails grade and if we need to make an update, they just have to run rails grade again to get the updates and don't have to worry about some funky command that we make up each time.