Add rake task to reset access token (#55)
Sometimes a student will get into a situation where they have the wrong access token and we need to reset it. Prior to this commit, resetting the token is not easy to do. There aren't simple steps we can instruct students to follow, and usually we need to resolve this in person.

This commit adds a rake task that will reset the Grades access token, which will simplify the process of helping a student reset their token.