Add non-Rails support (#66)
Previously, the ruby_grade_runner gem was created to handle non-Rails projects since grade_runner didn't add proper dependencies or require the right files. It has become too difficult to manage both versions of these gems.

This commit adds non-Rails support with the following changes:
* adds activesupport as a runtime dependency
* adds rake as a runtime dependency
* loads required gems correctly
* only loads Railite and migrates database when in Rails env
* bumps oj gem version to resolve Ruby 3 warning

An additional feature/bug fix in this commit is the ability to record the GitHub username of the person running the grade Task. Previously, the Task was just reporting the full name of the user from the local Git config. Now, an API request is made to GitHub to retrieve the username (via primary email) and fallback to the full name if one is not found.