Fail gracefully if parent repo doesn't exist or have specs (#67)
With the addition of #57, now rails grade will fail if the parent repo:
* Doesn't exist
* Doesn't have specs

Most of the time this happens when developing locally for a project that hasn't been published yet, but it's possible that it breaks if students fork a copy of a repo and name it differently.

This commit quits the task with a descriptive message if a project isn't available on appdev-projects or doesn't have a specs folder. This is to keep in line with the idea that the only specs that should be run are the ones available on appdev-projects.