Make git credentials optional (#76)
A recent Gitpod change has made it so Git credentials (user.email and user.name) are no longer automatically set.

This causes a problem because the email and username are used when submitting results to grades.firstdraft.com. Without them, the task fails.

This commit makes git credentials optional, allowing the grade task to complete even if the git credentials have not been set.