Unstage new spec file contents
Previously, unstaged and committed changes to the spec folder were removed before downloading the contents from the remote repo. This had the affect of staging the new spec file contents and preventing them from being removed is the token was updated later. This commit updates the logic related to syncing the remote spec files to unstage the new file contents. This allows the process of updating the grade token to work indefinitely while preserving the current behavior of discarding unstaged and committed changes. The other approach of committing the new spec folder contents was not implemented to avoid confusing learning with a git history that they did not explicitly create.