Remove deprecated rspec order flag (#81)
After project syncing on https://github.com/firstdraft/project-syncing/pull/4, I found that there was still the deprecated `--order default` flag directly on the rspec call here. This needs to be removed or the `grade` command will fail due to the updated RSpec version in all projects.

The version of grade_runner is also bumped here.