Sync project specs with Grades endpoint (#86)
Previously, the grade task would make 3 unauthenticated GitHub API calls per execution. This made it possible for students to reach the API request limit of 60 per hour if they ran the tests at least 20 times in an hour.
 
This commit makes the following changes:
- grade runner now retrieves spec folder sha and URL to project source code from Grades API endpoint
- when remote spec folder sha does not match local spec folder sha, grade runner now downloads a ZIP of the source code for the project, unzips it and overwrites the local spec folder
- when specs are updated, grade runner now makes a commit of the spec folder changes to ensure that the updated local spec folder sha will match the remote spec folder sha

Additionally, the grade task now retrieves and saves the GitHub username to the YAML file. This is still an unauthenticated call, but it should only happen once per project.