Allow configuration of grading behavior (#88)
In AD2, we'd like students to be able to write specs but also be able to run rake grade and get points.

We discussed a flow:
- projects start with a small number of specs implemented (maybe 0)
- all other specs have been stubbed out (but have points)
- students need to implement them to get points

With projects starting with all total specs, the grade percentage calculation can still happen normally.
Additionally, we add another configuration setting to default all specs to be worth 1 (or some other specified number of) points. This could be an issue if a student adds additional specs that aren't part of the starting stubbed out ones. For AD2 students, I don't this this is a big issue.

Crucially, for this to work grade_runner cannot overwrite contents of spec folder.

This commit adds:
- default formatters for specs and hints
  - these can be overridden by formatters present in the Ruby project
- configuration for default points for each spec
- configuration to allow/disallow overriding local specs